# specs/1-autonomous-business-agent/contracts/rest-api.openapi.yaml
openapi: 3.1.0
info:
  title: Autonomous Business Agent API
  version: 1.0.0
  description: |
    REST API for the Autonomous Business Agent system.
    Covers goal submission, plan approval/cancellation, and run management.
    Audit log queries are handled by the GraphQL endpoint (see graphql.schema.graphql).

servers:
  - url: /api/v1

security:
  - BearerAuth: []

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:

    Goal:
      type: object
      required: [id, text, submitted_by, submitted_at, status]
      properties:
        id:
          type: string
          format: uuid
        text:
          type: string
          maxLength: 2000
          description: Natural-language business objective
        submitted_by:
          type: string
          format: uuid
        submitted_at:
          type: string
          format: date-time
        status:
          type: string
          enum: [received, planning, plan_ready, approved, executing, completed, cancelled, failed]

    Plan:
      type: object
      required: [id, goal_id, created_at, status, tasks]
      properties:
        id:
          type: string
          format: uuid
        goal_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        status:
          type: string
          enum: [draft, approved, cancelled]
        tasks:
          type: array
          minItems: 1
          maxItems: 20
          items:
            $ref: '#/components/schemas/Task'

    Task:
      type: object
      required: [id, plan_id, description, dependencies, complexity, type, timeout_seconds, order_index, status, is_critical_path]
      properties:
        id:
          type: string
          format: uuid
        plan_id:
          type: string
          format: uuid
        description:
          type: string
        dependencies:
          type: array
          items:
            type: string
            format: uuid
        complexity:
          type: string
          enum: [low, medium, high]
        type:
          type: string
          enum: [data_retrieval, content_generation, notification_dispatch]
        timeout_seconds:
          type: integer
          default: 60
        order_index:
          type: integer
        status:
          type: string
          enum: [queued, running, succeeded, failed, skipped, cancelled]
        is_critical_path:
          type: boolean

    ExecutionRun:
      type: object
      required: [id, plan_id, triggered_by, started_at, status, tasks_total, tasks_succeeded, tasks_failed, tasks_skipped]
      properties:
        id:
          type: string
          format: uuid
        plan_id:
          type: string
          format: uuid
        triggered_by:
          type: string
          format: uuid
        started_at:
          type: string
          format: date-time
        ended_at:
          type: string
          format: date-time
          nullable: true
        status:
          type: string
          enum: [pending, running, completed, failed, cancelled]
        auto_execute:
          type: boolean
        tasks_total:
          type: integer
        tasks_succeeded:
          type: integer
        tasks_failed:
          type: integer
        tasks_skipped:
          type: integer

    LogEntry:
      type: object
      required: [id, run_id, timestamp, status, reasoning_summary, input_snapshot, output_snapshot, retry_count]
      properties:
        id:
          type: string
          format: uuid
        run_id:
          type: string
          format: uuid
        task_id:
          type: string
          format: uuid
          nullable: true
        timestamp:
          type: string
          format: date-time
          description: ISO 8601 timestamp
        status:
          type: string
          enum: [queued, running, succeeded, failed, skipped, cancelled, retrying]
        reasoning_summary:
          type: string
        input_snapshot:
          type: object
        output_snapshot:
          type: object
        error_context:
          type: string
          nullable: true
        retry_count:
          type: integer

    RunStatusEvent:
      type: object
      description: SSE event payload for real-time run status
      required: [event, run_id, timestamp]
      properties:
        event:
          type: string
          enum: [task_queued, task_started, task_succeeded, task_failed, task_skipped, task_retrying, run_completed, run_failed, run_cancelled]
        run_id:
          type: string
          format: uuid
        task_id:
          type: string
          format: uuid
          nullable: true
        timestamp:
          type: string
          format: date-time
        data:
          type: object
          description: Event-specific payload

    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object

paths:

  # ── Goals ──────────────────────────────────────────────

  /goals:
    post:
      operationId: submitGoal
      summary: Submit a new business goal
      description: |
        Accepts a natural-language business goal. The agent begins planning immediately.
        Returns the created goal. Poll GET /goals/{id} or GET /goals/{id}/plan for status.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text:
                  type: string
                  maxLength: 2000
                  description: Natural-language business objective
            example:
              text: "Generate a weekly sales summary and send it to the leadership team"
      responses:
        '202':
          description: Goal accepted, planning initiated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Goal'
        '400':
          description: Invalid input (e.g., text exceeds 2000 chars or empty)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Duplicate goal — an active run already exists for this goal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated

  /goals/{goalId}:
    get:
      operationId: getGoal
      summary: Get goal details and current status
      parameters:
        - name: goalId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Goal details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Goal'
        '404':
          description: Goal not found

  /goals/{goalId}/plan:
    get:
      operationId: getGoalPlan
      summary: Get the generated plan for a goal
      description: Returns the plan once planning is complete (goal.status = plan_ready or later).
      parameters:
        - name: goalId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Plan with tasks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Plan'
        '202':
          description: Plan not yet ready — agent is still planning
        '404':
          description: Goal not found

  /goals/{goalId}/plan/approve:
    post:
      operationId: approvePlan
      summary: Approve the generated plan and start execution
      description: |
        Transitions the plan from draft → approved and creates an ExecutionRun.
        Only available when goal.status = plan_ready.
        If the operator has auto_execute=true, this endpoint is called automatically.
      parameters:
        - name: goalId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '201':
          description: Run created and execution started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionRun'
        '409':
          description: Plan not in draft state or run already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Goal or plan not found

  /goals/{goalId}/plan/cancel:
    post:
      operationId: cancelPlan
      summary: Cancel the plan before execution begins
      parameters:
        - name: goalId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Plan cancelled, logged
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Plan'
        '409':
          description: Plan already approved or execution already started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ── Runs ───────────────────────────────────────────────

  /runs:
    get:
      operationId: listRuns
      summary: List execution runs
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, running, completed, failed, cancelled]
        - name: triggered_by
          in: query
          schema:
            type: string
            format: uuid
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Paginated list of runs
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/ExecutionRun'
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer

  /runs/{runId}:
    get:
      operationId: getRun
      summary: Get execution run details and summary
      parameters:
        - name: runId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Run details with task summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionRun'
        '404':
          description: Run not found

  /runs/{runId}/cancel:
    post:
      operationId: cancelRun
      summary: Cancel an in-progress execution run
      description: Cancellation is logged as a log entry. In-flight tasks are allowed to finish their current step before stopping.
      parameters:
        - name: runId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Cancellation initiated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionRun'
        '409':
          description: Run is not in a cancellable state

  /runs/{runId}/status/stream:
    get:
      operationId: streamRunStatus
      summary: Subscribe to real-time run status via Server-Sent Events
      description: |
        Returns a persistent SSE stream. Each event is a JSON-encoded RunStatusEvent.
        The stream closes when the run reaches a terminal state (completed, failed, cancelled).
        Clients must send the standard `Accept: text/event-stream` header.
      parameters:
        - name: runId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: SSE stream of run status events
          content:
            text/event-stream:
              schema:
                $ref: '#/components/schemas/RunStatusEvent'

  /runs/{runId}/logs:
    get:
      operationId: getRunLogs
      summary: Get execution logs for a run
      description: Returns log entries in chronological order. For advanced filtering use the GraphQL endpoint.
      parameters:
        - name: runId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: task_id
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            type: string
            enum: [queued, running, succeeded, failed, skipped, cancelled, retrying]
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
            maximum: 500
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Paginated log entries
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/LogEntry'
                  total:
                    type: integer

  # ── Auth ────────────────────────────────────────────────

  /auth/token:
    post:
      operationId: getToken
      summary: Obtain a JWT access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Access and refresh tokens
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  refresh_token:
                    type: string
                  expires_in:
                    type: integer
                    description: Seconds until access_token expires
        '401':
          description: Invalid credentials

  /auth/refresh:
    post:
      operationId: refreshToken
      summary: Refresh an expired access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
      responses:
        '200':
          description: New access token
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  expires_in:
                    type: integer
        '401':
          description: Invalid or expired refresh token